Move NSError to the step package
